+++ src/lib/types/domain/index.ts
@@ -15,4 +15,5 @@ export * from './user';
 export * from './banner-notification';
 export * from './purchase';
 export * from './visitor-identity';
+export * from './sms-blast';
 export * from './video';